+++ src/main/java/insane96mcp/iguanatweaksreborn/modules/misc/feature/TempSpawnerFeature.java
@@ -7,7 +7,10 @@ import insane96mcp.iguanatweaksreborn.utils.LogHelper;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.config.BlacklistConfig;
+import insane96mcp.insanelib.utils.IdTagMatcher;
 import net.minecraft.block.Blocks;
+import net.minecraft.entity.EntityType;
 import net.minecraft.entity.SpawnReason;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.particles.ParticleTypes;
@@ -20,30 +23,37 @@ import net.minecraftforge.event.entity.living.LivingSpawnEvent;
 import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+
 @Label(name = "Temporary Spawners", description = "Spawners will no longer spawn mobs infinitely")
 public class TempSpawnerFeature extends Feature {
 
     private final ForgeConfigSpec.ConfigValue<Integer> minSpawnableMobsConfig;
     private final ForgeConfigSpec.ConfigValue<Double> spawnableMobsMultiplierConfig;
-    //TODO Add a mob + dimension blacklist (dimension optional)
     private final ForgeConfigSpec.ConfigValue<Boolean> bonusExperienceWhenFarFromSpawnConfig;
+    private final BlacklistConfig entityBlacklistConfig;
 
     public int minSpawnableMobs = 25;
     public double spawnableMobsMultiplier = 1.0d;
     public boolean bonusExperienceWhenFarFromSpawn = true;
+    public List<IdTagMatcher> entityBlacklist;
+    public boolean entityBlacklistAsWhitelist = false;
 
     public TempSpawnerFeature(Module module) {
         super(Config.builder, module);
         Config.builder.comment(this.getDescription()).push(this.getName());
         minSpawnableMobsConfig = Config.builder
                 .comment("The minimum amount of spawnable mobs (when the spawner is basically in the same position as the world spawn. The amount of spawnable mobs before deactivating is equal to the distance divided by 8 (plus this value). E.g. At 160 blocks from spawn the max spawnable mobs will be 160 / 8 + 25 = 20 + 25 = 55")
-                .defineInRange("Spawnable mobs multiplier", minSpawnableMobs, 0, Integer.MAX_VALUE);
+                .defineInRange("Minimum Spawnable Mobs", minSpawnableMobs, 0, Integer.MAX_VALUE);
         spawnableMobsMultiplierConfig = Config.builder
                 .comment("This multiplier increases the max mobs spawned.")
                 .defineInRange("Spawnable mobs multiplier", spawnableMobsMultiplier, 0d, Double.MAX_VALUE);
         bonusExperienceWhenFarFromSpawnConfig = Config.builder
                 .comment("If true, the spawner will drop more experience when broken based of distance from spawn. +100% every 1024 blocks from spawn. The multiplier from 'Experience From Blocks' Feature still applies.")
                 .define("Bonus experience the farther from spawn", bonusExperienceWhenFarFromSpawn);
+        entityBlacklistConfig = new BlacklistConfig(Config.builder, "Entity Blacklist", "A list of mobs (and optionally dimensions) that shouldn't have their spawner disabled. Each entry has an entity or entity tag and optionally a dimension. E.g. [\"minecraft:zombie\", \"minecraft:blaze,minecraft:the_nether\"]", Collections.emptyList(), false);
         Config.builder.pop();
     }
 
@@ -53,6 +63,8 @@ public class TempSpawnerFeature extends Feature {
         this.minSpawnableMobs = this.minSpawnableMobsConfig.get();
         this.spawnableMobsMultiplier = this.spawnableMobsMultiplierConfig.get();
         this.bonusExperienceWhenFarFromSpawn = this.bonusExperienceWhenFarFromSpawnConfig.get();
+        this.entityBlacklist = IdTagMatcher.parseStringList(this.entityBlacklistConfig.listConfig.get());
+        this.entityBlacklistAsWhitelist = this.entityBlacklistConfig.listAsWhitelistConfig.get();
     }
 
     @SubscribeEvent
@@ -63,6 +75,8 @@ public class TempSpawnerFeature extends Feature {
             return;
         if (event.getSpawner() == null)
             return;
+        CompoundNBT nbt = new CompoundNBT();
+        event.getSpawner().write(nbt);
         BlockPos spawnerPos = event.getSpawner().getSpawnerPosition();
         ServerWorld world = (ServerWorld) event.getWorld();
         MobSpawnerTileEntity mobSpawner = (MobSpawnerTileEntity) world.getTileEntity(spawnerPos);
@@ -70,11 +84,26 @@ public class TempSpawnerFeature extends Feature {
         if (spawnerCap == null)
             LogHelper.error("Something's wrong. The spawner has no capability");
         spawnerCap.addSpawnedMobs(1);
+        //If it's in the black/whitelist don't disable the spawner
+        Optional<EntityType<?>> optional = EntityType.readEntityType(nbt.getCompound("SpawnData"));
+        if (!optional.isPresent())
+            return;
+        boolean isInWhitelist = false;
+        boolean isInBlacklist = false;
+        for (IdTagMatcher blacklistEntry : this.entityBlacklist) {
+            if (blacklistEntry.matchesEntity(optional.get())) {
+                if (!this.entityBlacklistAsWhitelist)
+                    isInBlacklist = true;
+                else
+                    isInWhitelist = true;
+                break;
+            }
+        }
+        if (isInBlacklist || (!isInWhitelist && this.entityBlacklistAsWhitelist))
+            return;
         double distance = Math.sqrt(spawnerPos.distanceSq(world.getSpawnPoint()));
         int maxSpawned = (int) ((this.minSpawnableMobs + (distance / 8d)) * this.spawnableMobsMultiplier);
         if (spawnerCap.getSpawnedMobs() >= maxSpawned) {
-            CompoundNBT nbt = new CompoundNBT();
-            event.getSpawner().write(nbt);
             nbt.putShort("MaxNearbyEntities", (short) 0);
             nbt.putShort("RequiredPlayerRange", (short) 0);
             event.getSpawner().read(nbt);
@@ -95,6 +124,17 @@ public class TempSpawnerFeature extends Feature {
     }
 
     public void onTick(MobSpawnerTileEntity spawner) {
+        //If the feature is disabled then reactivate disabled spawners and prevent further processing
+        if (!this.isEnabled()) {
+            CompoundNBT nbt = new CompoundNBT();
+            spawner.getSpawnerBaseLogic().write(nbt);
+            if (nbt.getShort("MaxNearbyEntities") == (short) 0 && nbt.getShort("RequiredPlayerRange") == (short) 0) {
+                nbt.putShort("MaxNearbyEntities", (short) 6);
+                nbt.putShort("RequiredPlayerRange", (short) 16);
+                spawner.getSpawnerBaseLogic().read(nbt);
+            }
+            return;
+        }
         World world = spawner.getWorld();
         CompoundNBT nbt = new CompoundNBT();
         spawner.write(nbt);